refactor: refactored audio and transcript function into component
resolves #360 , #361
